[PATCH] Output MySQL errno to result status
From: Nala Ginrut <nalaginrut@gmail.com>
Here's a patch to output MySQL errno to the result status.
This is important for users. If we don't export it, then user can
hardly detect the status os MySQL correctly.

For example, if the connect is timeout, then the status will be:
(2006 . MySQL server has gone away)
The number 2006 was defined as CR_SERVER_GONE_ERROR

And if I typed a wrong SQL statement, the status will be:
(1064 . You have an error in your SQL syntax; check the manual that
corresponds to your MariaDB server version for the right syntax to use
near 'null' at line 1)

In the current implementation of guile-dbd-mysql, the errno is always
1, which makes me hard to detect the status correctly, unless I parse
the reason string.